+++ test/spec/itemset.spec.js
@@ -333,6 +333,48 @@ describe( 'Itemset functionality', () => {
         } );
     } );
 
+    describe( 'in another group that becomes relevant (different from previous)', () => {
+        it( 'are re-evaluated', () => {
+            const form = loadForm( 'itemset-relevant-2.xml' );
+            form.init();
+            const input1 = form.view.html.querySelector( '[name="/data/start"]' );
+            input1.value = 'A';
+            input1.dispatchEvent( events.Change() );
+            const input2 = form.view.html.querySelector( '[name="/data/proceed"]' );
+            input2.checked = true;
+            input2.dispatchEvent( events.Change() );
+            const EXPECTED = [ 'AK', 'AL', 'AR', 'AZ' ];
+
+            // group with relevant
+            expect( [ ...form.view.html.querySelectorAll( '.option-wrapper label:not(.itemset-template) input[data-name="/data/grp1/one"]' ) ]
+                .map( input => input.value ) ).toEqual( EXPECTED );
+            expect( [ ...form.view.html.querySelectorAll( '.option-wrapper label:not(.itemset-template) input[name="/data/grp1/two"]' ) ]
+                .map( input => input.value ) ).toEqual( EXPECTED );
+            expect( [ ...form.view.html.querySelectorAll( 'select[name="/data/grp1/three"] option:not(.itemset-template)' ) ]
+                .map( input => input.value ) ).toEqual( EXPECTED );
+            expect( [ ...form.view.html.querySelectorAll( 'select[name="/data/grp1/four"] option:not(.itemset-template)' ) ]
+                .map( input => input.value ) ).toEqual( EXPECTED );
+            expect( [ ...form.view.html.querySelectorAll( 'input[name="/data/grp1/five"] ~ datalist option:not(.itemset-template)' ) ]
+                .map( input => input.dataset.value ) ).toEqual( EXPECTED );
+
+            // shared static autocomplete datalist in repeat
+            expect( [ ...form.view.html.querySelectorAll( '.or-repeat-info datalist option:not(.itemset-template)' ) ]
+                .map( input => input.dataset.value ) ).toEqual( [ 'AL' ] );
+
+            // individual itemset question with relevants
+            expect( [ ...form.view.html.querySelectorAll( '.option-wrapper label:not(.itemset-template) input[data-name="/data/grp2/seven"]' ) ]
+                .map( input => input.value ) ).toEqual( EXPECTED );
+            expect( [ ...form.view.html.querySelectorAll( '.option-wrapper label:not(.itemset-template) input[name="/data/grp2/eight"]' ) ]
+                .map( input => input.value ) ).toEqual( EXPECTED );
+            expect( [ ...form.view.html.querySelectorAll( 'select[name="/data/grp2/nine"] option:not(.itemset-template)' ) ]
+                .map( input => input.value ) ).toEqual( EXPECTED );
+            expect( [ ...form.view.html.querySelectorAll( 'select[name="/data/grp2/ten"] option:not(.itemset-template)' ) ]
+                .map( input => input.value ) ).toEqual( EXPECTED );
+            expect( [ ...form.view.html.querySelectorAll( 'input[name="/data/grp2/eleven"] ~ datalist option:not(.itemset-template)' ) ]
+                .map( input => input.dataset.value ) ).toEqual( EXPECTED );
+        } );
+    } );
+
     describe( 'in a group that becomes relevant, irrelevant relevant', () => {
         it( 'are re-evaluated', () => {
             const form = loadForm( 'itemset-relevant.xml' );
@@ -404,7 +446,7 @@ describe( 'Itemset functionality', () => {
                         <item>
                             <label>11</label>
                             <group>a</group>
-                            <name>1 1</name>                        
+                            <name>1 1</name>
                         </item>
                         <item>
                             <label>22</label>
@@ -462,7 +504,7 @@ describe( 'Itemset functionality', () => {
             [ 'instance("countries")/root/item[43]', true ],
             [ 'instance("countries")/root/item[label=current()/../answer]', false ]
         ].forEach( test => {
-            it( `correctly determines that ${test[0]} is ${test[1] === true ? '': 'not '}dynamic`, () => {
+            it( `correctly determines that ${test[0]} is ${test[1] === true ? '' : 'not '}dynamic`, () => {
                 expect( isStaticItemsetFromSecondaryInstance( test[ 0 ] ) ).toEqual( test[ 1 ] );
             } );
         } );